requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
numpy>=1.24.0
scikit-learn>=1.2.2
gradio>=3.35.2
python-dotenv>=1.0.0
tensorboard>=2.13.0
tqdm>=4.65.0
matplotlib>=3.7.1
